+++ packages/react-ui/src/components/base/index.ts
@@ -1,0 +1,18 @@
+export {
+  Avatar,
+  AvatarRoot,
+  AvatarImage,
+  AvatarFallback,
+  type AvatarProps,
+} from "./avatar";
+
+export { Button, buttonVariants, type ButtonProps } from "./button";
+
+export { Tooltip, TooltipContent, TooltipTrigger } from "./tooltip";
+
+export {
+  TooltipIconButton,
+  type TooltipIconButtonProps,
+} from "./tooltip-icon-button";
+
+export { CircleStopIcon } from "./CircleStopIcon";