+++ packages/react-ui/src/index.ts
@@ -1,2 +1,1 @@
-export { Thread } from "../components/ui/assistant-ui/thread";
-export { AssistantModal } from "../components/ui/assistant-ui/assistant-modal";
+export * from "./components";