+++ packages/react-ui/tailwind.config.ts
@@ -15,7 +15,7 @@ const config = {
     },
     extend: {
       colors: {
-        border: "hsl(var(--aui-border))",
+        //   border: "hsl(var(--aui-border))",
         input: "hsl(var(--aui-input))",
         ring: "hsl(var(--aui-ring))",
         background: "hsl(var(--aui-background))",
@@ -24,14 +24,14 @@ const config = {
           DEFAULT: "hsl(var(--aui-primary))",
           foreground: "hsl(var(--aui-primary-foreground))",
         },
-        secondary: {
-          DEFAULT: "hsl(var(--aui-secondary))",
-          foreground: "hsl(var(--aui-secondary-foreground))",
-        },
-        destructive: {
-          DEFAULT: "hsl(var(--aui-destructive))",
-          foreground: "hsl(var(--aui-destructive-foreground))",
-        },
+        //   secondary: {
+        //     DEFAULT: "hsl(var(--aui-secondary))",
+        //     foreground: "hsl(var(--aui-secondary-foreground))",
+        //   },
+        //   destructive: {
+        //     DEFAULT: "hsl(var(--aui-destructive))",
+        //     foreground: "hsl(var(--aui-destructive-foreground))",
+        //   },
         muted: {
           DEFAULT: "hsl(var(--aui-muted))",
           foreground: "hsl(var(--aui-muted-foreground))",
@@ -44,30 +44,16 @@ const config = {
           DEFAULT: "hsl(var(--aui-popover))",
           foreground: "hsl(var(--aui-popover-foreground))",
         },
-        card: {
-          DEFAULT: "hsl(var(--aui-card))",
-          foreground: "hsl(var(--aui-card-foreground))",
-        },
+        //   card: {
+        //     DEFAULT: "hsl(var(--aui-card))",
+        //     foreground: "hsl(var(--aui-card-foreground))",
+        //   },
       },
-      borderRadius: {
-        lg: "var(--aui-radius)",
-        md: "calc(var(--aui-radius) - 2px)",
-        sm: "calc(var(--aui-radius) - 4px)",
-      },
-      keyframes: {
-        "accordion-down": {
-          from: { height: "0" },
-          to: { height: "var(--aui-radix-accordion-content-height)" },
-        },
-        "accordion-up": {
-          from: { height: "var(--aui-radix-accordion-content-height)" },
-          to: { height: "0" },
-        },
-      },
-      animation: {
-        "accordion-down": "accordion-down 0.2s ease-out",
-        "accordion-up": "accordion-up 0.2s ease-out",
-      },
+      // borderRadius: {
+      //   lg: "var(--aui-radius)",
+      //   md: "calc(var(--aui-radius) - 2px)",
+      //   sm: "calc(var(--aui-radius) - 4px)",
+      // },
     },
   },
   plugins: [require("tailwindcss-animate")],